requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-classic
langgraph
zhipuai
faiss-cpu
python-dotenv
pypdf
sqlite3